refactor(http-protocol): restructure scrape responses into layered modules
Flat scrape.rs and scrape_deserialization.rs replaced by scrape/{data,encoding,deserialization}.rs directory module.

- data.rs - DTO layer: SwarmMetadata, ScrapeData
- encoding.rs - Encoding layer: Bencoded struct + tests
- deserialization.rs - Client-side: Response, File, ResponseBuilder, BencodeParseError
- Deleted both old flat files
- Updated all 7 import sites
- All linters and tests pass
- Progress: T12-T14 done, spec updated
