+++ console/tracker-client/src/console/clients/http/mod.rs
@@ -2,7 +2,7 @@ use std::sync::Arc;
 
 use serde::Serialize;
 use thiserror::Error;
-use torrust_tracker_http_protocol::v1::responses::scrape_deserialization::BencodeParseError;
+use torrust_tracker_http_protocol::v1::responses::scrape::deserialization::BencodeParseError;
 
 pub mod app;
