+++ docs/issues/open/1965-1669-si-34-consolidate-duplicate-http-types/ISSUE.md
@@ -226,9 +226,9 @@ Status values: `TODO`, `IN_PROGRESS`, `BLOCKED`, `DONE`.
 |     |        | **Response-side restructuring (DD7 + DD8)**           |                                                                                                                                                                                                                                                                                                                                                 |
 | T10 | DONE   | Restructure announce responses into layered module    | Created `announce/{data,encoding}.rs` with `mod.rs` re-exports. Deleted old `announce.rs`. No call sites needed updating (backward compatible)                                                                                                                                                                                                  |
 | T11 | DONE   | Partial merge of announce DTO layer                   | Moved deserialization types into `announce/deserialization.rs`. Renamed `Announce` → `DeserializedNormal`, `Compact` → `DeserializedCompactParsed`. Added `peers6` to `DeserializedCompact`. Replaced `CompactPeer` (IPv4-only struct) with shared `encoding::CompactPeer` enum. Deleted `announce_deserialization.rs`. Updated 8 import sites. |
-| T12 | TODO   | Restructure scrape responses into layered module      | Same pattern: `scrape/{data,encoding,deserialization}.rs`                                                                                                                                                                                                                                                                                       |
-| T13 | TODO   | Partial merge of scrape DTO layer                     | Same pattern as announce                                                                                                                                                                                                                                                                                                                        |
-| T14 | TODO   | Update all call sites for restructured response types | Update imports in tracker-client, axum-http-server tests, CLI apps                                                                                                                                                                                                                                                                              |
+| T12 | DONE   | Restructure scrape responses into layered module      | Created `scrape/{data,encoding,deserialization}.rs`. Deleted `scrape.rs` and `scrape_deserialization.rs`. Updated 7 import sites. Backward compatible re-exports.                                                                                                                                                                               |
+| T13 | DONE   | Partial merge of scrape DTO layer                     | Merged `scrape_deserialization.rs` into `scrape/deserialization.rs`. Updated all import sites. Done together with T12.                                                                                                                                                                                                                          |
+| T14 | DONE   | Update all call sites for restructured response types | Updated all import sites for both announce and scrape restructuring. Done together with T10-T13.                                                                                                                                                                                                                                                |
 | T15 | TODO   | Run full verification after response restructure      | `linter all`, `cargo test --workspace`, pre-commit, pre-push                                                                                                                                                                                                                                                                                    |
 |     |        | **Finalization**                                      |                                                                                                                                                                                                                                                                                                                                                 |
 | T16 | TODO   | Replace duplicate HTTP test client (DD9)              | Remove `packages/axum-http-server/tests/server/client.rs`; update tests to use `tracker-client` package                                                                                                                                                                                                                                         |
@@ -260,6 +260,7 @@ Status values: `TODO`, `IN_PROGRESS`, `BLOCKED`, `DONE`.
 - 2026-07-14 10:00 UTC - Copilot - Implementation (T7-T9) completed: merged `announce_builder::Query` into `Announce`, updated all call sites, all verifications passed.
 - 2026-07-14 14:00 UTC - Copilot - Implementation (T10) completed: restructured announce responses into `announce/{data,encoding}.rs` layered module.
 - 2026-07-14 15:00 UTC - Copilot - Implementation (T11) completed: partial merge of announce DTO layer into `announce/deserialization.rs`.
+- 2026-07-14 16:00 UTC - Copilot - Implementation (T12-T14) completed: restructured scrape responses into `scrape/{data,encoding,deserialization}.rs`, merged DTO layer, updated all import sites.
 
 ## Acceptance Criteria
