+++ packages/http-protocol/src/v1/responses/mod.rs
@@ -2,6 +2,5 @@
 pub mod announce;
 pub mod error;
 pub mod scrape;
-pub mod scrape_deserialization;
 
 pub use announce::{Announce, Compact, Normal};
