+++ packages/http-protocol/src/v1/responses/scrape/deserialization.rs
