+++ packages/http-protocol/src/v1/responses/scrape/mod.rs
@@ -1,0 +1,8 @@
+//! Scrape response types for the HTTP tracker.
+pub mod data;
+pub mod deserialization;
+pub mod encoding;
+
+pub use data::{ScrapeData, SwarmMetadata};
+pub use deserialization::{BencodeParseError, File, Response, ResponseBuilder};
+pub use encoding::Bencoded;
